backend/requirements.txt
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
pymongo[srv]>=4.5.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-socketio>=5.8.0
eventlet>=0.33.0
requests>=2.31.0
python-dateutil>=2.8.0
marshmallow>=3.20.0
